add notification

+++ addon.py
@@ -23,24 +23,31 @@ for i in range(1, MAX_OS+1):
             "nb_part": nb_part
         })
 
-os_idx = xbmcgui.Dialog().select("Choose OS", [x["name"] for x in os_list])
+if len(os_list) > 0:
+    os_idx = xbmcgui.Dialog().select("Choose OS", [x["name"] for x in os_list])
 
-# check nb_part > 5 because partition 0 to 5 are reserved to pinn system
-if int(os_list[os_idx]["nb_part"]) <= 5:
-    xbmcgui.Dialog().ok("Starting Raspbian", "Error", "Os '%s' has wrong partition configuration, it has to be superior to 5 (partitions 1 to 5 are reserved to pinn system)." % os_list[os_idx]["name"])
+    # check nb_part > 5 because partition 0 to 5 are reserved to pinn system
+    if int(os_list[os_idx]["nb_part"]) <= 5:
+        xbmcgui.Dialog().ok(addonname, "Error", "Os '%s' has wrong partition configuration, it has to be superior to 5 (partitions 1 to 5 are reserved to pinn system)." % os_list[os_idx]["name"])
+    else:
+        reboot = False
+
+        try:
+            if os.path.isfile(autoboot_file):
+                os.remove(autoboot_file)
+
+            if os.path.isdir(pinn_sys_part_path):
+                with open(autoboot_file, "w") as f:
+                    f.write("boot_partition=%s" % os_list[os_idx]["nb_part"])
+                reboot = True
+        except Exception as e:
+            xbmcgui.Dialog().ok("Switching to %s" % os_list[os_idx]["name"], "Exception Error", str(e))
+
+        if reboot:
+            xbmcgui.Dialog().notification("Switching to %s" % os_list[os_idx]["name"], "please wait...", time=20000)
+            # import time
+            # with open("d:\\test.txt", "w+") as f:
+            #     f.write(time.strftime("%Y%m%d_%B%M%S"))
+            os.system("reboot")
 else:
-    reboot = False
-
-    try:
-        if os.path.isfile(autoboot_file):
-            os.remove(autoboot_file)
-
-        if os.path.isdir(pinn_sys_part_path):
-            with open(autoboot_file, "w") as f:
-                f.write("boot_partition=%s" % os_list[os_idx]["nb_part"])
-            reboot = True
-    except Exception as e:
-        xbmcgui.Dialog().ok("Starting Raspbian", "Error", str(e))
-
-    if reboot:
-        os.system("reboot")
+    xbmcgui.Dialog().ok(addonname, "Error", "No OS configured yet.")
